review rating summary graph edited

+++ lib/ui/widgets/minor_widgets/rating_summary_progress_bar.dart
@@ -15,8 +15,8 @@ class RatingSummaryProgressBar extends StatelessWidget {
   });
 
   final String text;
-  final double ratingCount;
-  final double totalRatingCount;
+  final int ratingCount;
+  final int totalRatingCount;
   final double? paddingHorizontal;
   final double? height;
